requirements.txt
steam[client]
GitPython
requests
openpyxl
tqdm
beautifulsoup4
